refactor: organize related code in folders
