+++ src/main/kotlin/com/atauchi/transformerFileService/core/domain/useCases/transformerFile/TransformerFileSaveUseCase.kt
@@ -1,4 +1,4 @@
-package com.atauchi.transformerFileService.core.domain.useCases
+package com.atauchi.transformerFileService.core.domain.useCases.transformerFile
 
 import com.atauchi.transformerFileService.core.domain.entities.transformerFile.User
 import org.springframework.web.multipart.MultipartFile
